+++ benches/resampler.rs
@@ -1,9 +1,11 @@
 use divan::Bencher;
+use rodio::decoder::DecoderSample;
 use rodio::source::UniformSourceIterator;
 
 mod shared;
+use shared::music_wav;
+
 use rodio::Source;
-use shared::TestSource;
 
 fn main() {
     divan::main();
@@ -13,11 +15,11 @@ fn main() {
 fn no_resampling(bencher: Bencher) {
     bencher
         .with_inputs(|| {
-            let source = TestSource::<i16>::music_wav();
+            let source = music_wav();
             (source.channels(), source.sample_rate(), source)
         })
         .bench_values(|(channels, sample_rate, source)| {
-            UniformSourceIterator::<_, i16>::new(source, channels, sample_rate)
+            UniformSourceIterator::<_, DecoderSample>::new(source, channels, sample_rate)
                 .for_each(divan::black_box_drop)
         })
 }
@@ -32,11 +34,11 @@ const COMMON_SAMPLE_RATES: [u32; 12] = [
 fn resample_to(bencher: Bencher, target_sample_rate: u32) {
     bencher
         .with_inputs(|| {
-            let source = TestSource::<i16>::music_wav();
+            let source = music_wav();
             (source.channels(), source)
         })
         .bench_values(|(channels, source)| {
-            UniformSourceIterator::<_, i16>::new(source, channels, target_sample_rate)
+            UniformSourceIterator::<_, DecoderSample>::new(source, channels, target_sample_rate)
                 .for_each(divan::black_box_drop)
         })
 }