+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `GeneratorFunction`.
 - Minimal builds without `cpal` audio output are now supported.
   See `README.md` for instructions. (#349)
+- Added `Sample::is_zero()` method for checking zero samples.
 
 ### Changed
 - Breaking: `OutputStreamBuilder` should now be used to initialize an audio output stream.
@@ -26,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Breaking: `Sink::try_new` renamed to `connect_new` and does not return error anymore.
             `Sink::new_idle` was renamed to `new`.
 - Breaking: In the `Source` trait, the method `current_frame_len()` was renamed to `current_span_len()`.
+- Breaking: `Decoder` now outputs `f32` samples by default instead of `i16`.
+  Enable the `integer-decoder` to revert to `i16` samples.
 - The term 'frame' was renamed to 'span' in the crate and documentation.
 
 ### Fixed
@@ -34,6 +37,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Symphonia decoder `total_duration` incorrect value caused by conversion from `Time` to `Duration`.
 - An issue with `SignalGenerator` that caused it to create increasingly distorted waveforms
   over long run times has been corrected. (#201)
+- WAV and FLAC decoder duration calculation now calculated once and handles very large files
+  correctly
+- Removed unwrap() calls in MP3, WAV, FLAC and Vorbis format detection for better error handling
+
+### Deprecated
+- Deprecated `Sample::zero_value()` function in favor of `Sample::ZERO_VALUE` constant
 
 # Version 0.20.1 (2024-11-08)